src/ml_pipeline/requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.2
python-dotenv==1.0.1
scikit-learn==1.5.0
numpy==1.26.4
joblib==1.4.2
loguru==0.7.2
fastapi==0.110.0
uvicorn[standard]==0.30.1
